refactor: :recycle: Remove old download manager and fix import paths everywhere
