--- boosty_downloader/src/application/download_manager/download_manager_config.py
@@ -1,75 +1,0 @@
-"""All necessary dependency containers for the main class"""
-
-from dataclasses import dataclass
-from enum import Enum
-from pathlib import Path
-
-from aiohttp_retry import RetryClient
-
-from boosty_downloader.src.infrastructure.boosty_api.core.client import BoostyAPIClient
-from boosty_downloader.src.infrastructure.boosty_api.models.post.post_data_types.post_data_ok_video import (
-    BoostyOkVideoType,
-)
-from boosty_downloader.src.infrastructure.external_videos_downloader.external_videos_downloader import (
-    ExternalVideosDownloader,
-)
-from boosty_downloader.src.infrastructure.loggers.base import RichLogger
-from boosty_downloader.src.infrastructure.loggers.failed_downloads_logger import (
-    FailedDownloadsLogger,
-)
-
-
-@dataclass
-class LoggerDependencies:
-    """Class that holds loggers for the download manager"""
-
-    failed_downloads_logger: FailedDownloadsLogger
-    logger: RichLogger
-
-
-@dataclass
-class NetworkDependencies:
-    """Class that holds network dependencies for the download manager"""
-
-    session: RetryClient
-    api_client: BoostyAPIClient
-    external_videos_downloader: ExternalVideosDownloader
-
-
-class DownloadContentTypeFilter(Enum):
-    """Class that holds content type filters for the download manager (such as videos, images, etc)"""
-
-    boosty_videos = 'boosty_videos'
-    external_videos = 'external_videos'
-    post_content = 'post_content'
-    files = 'files'
-
-
-class VideoQualityOption(str, Enum):
-    """Preferred video quality option for cli"""
-
-    smallest_size = 'smallest_size'
-    low = 'low'
-    medium = 'medium'
-    high = 'high'
-    highest = 'highest'
-
-    def to_ok_video_type(self) -> BoostyOkVideoType:
-        mapping: dict[VideoQualityOption, BoostyOkVideoType] = {
-            VideoQualityOption.smallest_size: BoostyOkVideoType.lowest,
-            VideoQualityOption.low: BoostyOkVideoType.low,
-            VideoQualityOption.medium: BoostyOkVideoType.medium,
-            VideoQualityOption.high: BoostyOkVideoType.high,
-            VideoQualityOption.highest: BoostyOkVideoType.ultra_hd,
-        }
-        return mapping[self]
-
-
-@dataclass
-class GeneralOptions:
-    """Class that holds general options for the download manager (such as paths)"""
-
-    target_directory: Path
-    download_content_type_filters: list[DownloadContentTypeFilter]
-    request_delay_seconds: float
-    preferred_video_quality: VideoQualityOption
