+++ boosty_downloader/src/application/filtering.py
@@ -2,6 +2,10 @@
 
 from enum import Enum
 
+from boosty_downloader.src.infrastructure.boosty_api.models.post.post_data_types.post_data_ok_video import (
+    BoostyOkVideoType,
+)
+
 
 class DownloadContentTypeFilter(Enum):
     """
@@ -14,3 +18,23 @@ class DownloadContentTypeFilter(Enum):
     external_videos = 'external_videos'
     post_content = 'post_content'
     files = 'files'
+
+
+class VideoQualityOption(str, Enum):
+    """Preferred video quality option for cli"""
+
+    smallest_size = 'smallest_size'
+    low = 'low'
+    medium = 'medium'
+    high = 'high'
+    highest = 'highest'
+
+    def to_ok_video_type(self) -> BoostyOkVideoType:
+        mapping = {
+            VideoQualityOption.smallest_size: BoostyOkVideoType.lowest,
+            VideoQualityOption.low: BoostyOkVideoType.low,
+            VideoQualityOption.medium: BoostyOkVideoType.medium,
+            VideoQualityOption.high: BoostyOkVideoType.high,
+            VideoQualityOption.highest: BoostyOkVideoType.ultra_hd,
+        }
+        return mapping[self]
